refactor: Game 관련 클래스 패키지 이동
- 좀 더 명확한 game 패키지 하위로 이동
