+++ src/test/java/com/wootecam/chess/game/CoordinatesExtractorTest.java
@@ -1,4 +1,4 @@
-package com.wootecam.chess;
+package com.wootecam.chess.game;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
