+++ src/test/java/com/wootecam/chess/game/PieceMoveVerifierTest.java
@@ -1,4 +1,4 @@
-package com.wootecam.chess;
+package com.wootecam.chess.game;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
